+++ src/main/java/org/realityforge/gwt/webpoller/client/WebPoller.java
@@ -1,6 +1,5 @@
 package org.realityforge.gwt.webpoller.client;
 
-import com.google.gwt.core.shared.GWT;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -32,6 +31,21 @@ public abstract class WebPoller
 
   private static Factory g_factory;
 
+  static
+  {
+    try
+    {
+      if ( (Boolean) Class.forName( "com.google.gwt.core.shared.GWT" ).getMethod( "isClient" ).invoke( null ) )
+      {
+        register( new TimerBasedWebPoller.Factory() );
+      }
+    }
+    catch ( final Throwable e )
+    {
+      //Ignored
+    }
+  }
+
   private final RequestContext _requestContext = new WebPollerRequestContext();
   private WebPollerListener _listener = NullWebPollerListener.INSTANCE;
   private RequestFactory _requestFactory;
@@ -49,10 +63,6 @@ public abstract class WebPoller
 
   public static WebPoller newWebPoller()
   {
-    if ( null == g_factory && GWT.isClient() )
-    {
-      register( new TimerBasedWebPoller.Factory() );
-    }
     return ( null != g_factory ) ? g_factory.newWebPoller() : null;
   }